Complete rewrite
Major updates throughout. The Black Panda is now stable and the requirements for Black vs White or Gray Panda users have changed. It is preferable to do the install via a physical Ethernet cable first and configure the WiFi later. The git clone is now one step. The install command is now one step. Flashing the Panda is completely changed now. The flash step is the same for Black and White or Gray Pandas, but owners of White or Gray Pandas must go through some extra setup and teardown steps when flashing. There is also a new section on the first drive and training session.