new one-press buttons
describe the new buttons that increase or decrease specific settings by the amount Gernby normally changes them during a live tuning session. this reduces driver interaction to a single button press to execute each change. this new approach makes it very easy for the driver to make single, specific changes and see how the car feels before moving on to the next change (if any)