create raspilot/models; delete buttons
not sure what happened recently but the ~/raspilot/models directory was no longer being created on new builds. Several testers are reporting that the folder doesn't exist when trying to deploy the model. I believe this is related to recent changes in gitignore

also, the button folders for model switching is deprecated and will be removed at a future date