clarified successful flash again
A successful flash should result in a slow, pulsing LED of one or two colors, and those colors change over time as new updates are published. The only thing you do not want to see after a flash attempt is a fast green blink. That means the attempt was not successful and you need to try again.